server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
google-genai==1.14.0
slack-bolt==1.22.0
slack-sdk==3.34.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
